購物車調整

+++ src/stores/cartStore.js
@@ -1,8 +1,7 @@
 const { VITE_URL, VITE_PATH } = import.meta.env
 import { defineStore } from 'pinia'
+import { Toast, Alert } from '@/mixins/swal'
 import axios from 'axios'
-import Toast from '@/mixins/toast.js'
-// import Swal from 'sweetalert2/dist/sweetalert2.js'
 
 export default defineStore('cart', {
   state: () => {
@@ -20,20 +19,17 @@ export default defineStore('cart', {
       const url = `${VITE_URL}/api/${VITE_PATH}/cart`
       axios.get(url)
         .then(res => {
-          console.log('取得購物車列表:', res)
+          console.log('取得購物車列表', res)
           this.cartNum = res.data.data.carts.reduce((pre, item) => pre + item.qty, 0)
           this.cart = res.data.data
           this.total = res.data.data.total
           this.final_total = res.data.data.final_total
         })
         .catch(err => {
-          console.log(err)
-          // Swal.fire({
-          //   icon: 'error',
-          //   title: '請確認',
-          //   text: 'Something went wrong!',
-          //   footer: '<a href="">Why do I have this issue?</a>'
-          // })
+          console.log('取得購物車列表失敗', err)
+          Alert.fire({
+            title: err.response.data.message,
+          })
         })
     },
 
@@ -50,15 +46,18 @@ export default defineStore('cart', {
       axios.post(url, data)
         .then(res => {
           console.log('加入購物車', res)
-          this.loadingItem = ''
           this.getCart()
+          this.loadingItem = ''
           Toast.fire({
             icon: 'success',
             title: '成功加入購物車'
           })
         })
         .catch(err => {
-          console.log(err)
+          console.log('加入購物車失敗', err)
+          Alert.fire({
+            title: err.response.data.message,
+          })
         })
     },
 
@@ -72,22 +71,40 @@ export default defineStore('cart', {
       }
       axios.put(url, { data })
         .then(res => {
-          console.log('更新數量:', res)
-          this.loadingItem = ''
+          console.log('更新數量', res)
           this.getCart()
+          this.loadingItem = ''
+          Toast.fire({
+            icon: 'success',
+            title: '成功修改數量'
+          })
+        })
+        .catch(err => {
+          console.log('更新數量失敗', err)
+          Alert.fire({
+            title: err.response.data.message,
+          })
         })
     },
 
     // 刪除購物車單一品項
     removeCartItem(id) {
+      this.loadingItem = id
       const url = `${VITE_URL}/api/${VITE_PATH}/cart/${id}`
       axios.delete(url)
         .then(res => {
-          console.log(res)
+          console.log('刪除單一產品', res)
           this.getCart()
+          this.loadingItem = ''
           Toast.fire({
             icon: 'success',
-            title: '成功刪除'
+            title: '成功刪除商品'
+          })
+        })
+        .catch(err => {
+          console.log('刪除單一產品失敗', err)
+          Alert.fire({
+            title: err.response.data.message,
           })
         })
     },
@@ -95,10 +112,27 @@ export default defineStore('cart', {
     // 刪除全部購物車
     deleteAllCart() {
       const url = `${VITE_URL}/api/${VITE_PATH}/cart`
-      axios.delete(url)
-        .then(res => {
-          console.log('購物車已清空', res)
+      if (this.cartNum !== 0) {
+        axios.delete(url)
+          .then(res => {
+            console.log('購物車已清空', res)
+            this.getCart()
+            Toast.fire({
+              icon: 'success',
+              title: '已刪除全部購物車'
+            })
+          })
+          .catch(err => {
+            console.log('購物車清空失敗', err)
+            Alert.fire({
+              title: err.response.data.message,
+            })
+          })
+      } else {
+        Alert.fire({
+          title: '購物車尚無任何產品',
         })
+      }
     },
 
     // 複製優惠券
